upload-effects module added
